Change API to create motions (again)
