+++ src/ColonyNetwork/Colony.ts
@@ -4,10 +4,6 @@ import {
   SignerOrProvider,
   Id,
   Extension,
-  ColonyRole,
-  getPermissionProofs,
-  getExtensionPermissionProofs,
-  getChildIndex,
 } from '@colony/colony-js';
 
 import {
@@ -17,7 +13,6 @@ import {
   DomainAdded_uint256_EventObject,
   DomainMetadataEventObject,
   FundingPotAddedEventObject,
-  MotionCreatedEventObject,
   OneTxPaymentMadeEventObject,
 } from '@colony/colony-js/extras';
 import type {
@@ -26,17 +21,17 @@ import type {
   ContractTransaction,
 } from 'ethers';
 
-import { ColonyToken } from './ColonyToken';
-import { extractEvent } from '../utils';
-import { ColonyNetwork } from './ColonyNetwork';
 import { MetadataKey, MetadataValue } from '../events';
-import { ParametersFrom2 } from '../types';
+import { extractEvent } from '../utils';
+import { ColonyToken } from './ColonyToken';
+import { ColonyNetwork } from './ColonyNetwork';
+import { VotingReputation } from './VotingReputation';
 
 export type SupportedColonyClient = ColonyClientV8 | ColonyClientV9;
-
-type SupportedColonyMethods = SupportedColonyClient['functions'];
-
-const SUPPORTED_VOTING_REPUTATION_VERSION = 4;
+export type SupportedColonyMethods = SupportedColonyClient['functions'];
+export interface SupportedExtensions {
+  motions?: VotingReputation;
+}
 
 export class Colony {
   /** The currently supported Colony version. If a Colony is not on this version it has to be upgraded.
@@ -51,6 +46,8 @@ export class Colony {
   address: string;
 
   colonyNetwork: ColonyNetwork;
+
+  ext: SupportedExtensions;
 
   version: number;
 
@@ -68,10 +65,12 @@ export class Colony {
   constructor(
     colonyNetwork: ColonyNetwork,
     colonyClient: SupportedColonyClient,
+    extensions: SupportedExtensions,
   ) {
     this.colonyClient = colonyClient;
     this.signerOrProvider = colonyClient.signer || colonyClient.provider;
     this.address = colonyClient.address;
+    this.ext = extensions;
     this.colonyNetwork = colonyNetwork;
     this.version = colonyClient.clientVersion;
   }
@@ -98,61 +97,6 @@ export class Colony {
     }
 
     return [data, receipt];
-  }
-
-  private async getVotingReputationClient() {
-    const votingReputationClient = await this.colonyClient.getExtensionClient(
-      Extension.VotingReputation,
-    );
-
-    if (
-      votingReputationClient.clientVersion !==
-      SUPPORTED_VOTING_REPUTATION_VERSION
-    ) {
-      throw new Error(
-        `The installed version ${votingReputationClient.clientVersion} of the VotingReputation extension is not supported. Please upgrade the extension in your Colony`,
-      );
-    }
-
-    return votingReputationClient;
-  }
-
-  private async createMotion<F extends keyof SupportedColonyMethods>(
-    signature: F,
-    actionDomain: BigNumberish,
-    actionRole: ColonyRole,
-    motionDomain: BigNumberish,
-    args: ParametersFrom2<SupportedColonyMethods[F]>,
-  ) {
-    const votingReputationClient = await this.getVotingReputationClient();
-
-    const [permissionDomainId, childSkillIndex] = await getPermissionProofs(
-      this.colonyClient,
-      actionDomain,
-      actionRole,
-      votingReputationClient.address,
-    );
-
-    const encodedAction = this.colonyClient.interface.encodeFunctionData(
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore I don't think this can be typed properly here
-      signature,
-      [permissionDomainId, childSkillIndex, ...args],
-    ) as string;
-
-    const tx = await votingReputationClient.createMotionWithProofs(
-      motionDomain,
-      '0x0',
-      encodedAction,
-    );
-
-    const receipt = await tx.wait();
-
-    const data = {
-      ...extractEvent<MotionCreatedEventObject>('MotionCreated', receipt),
-    };
-
-    return [data, receipt] as [typeof data, typeof receipt];
   }
 
   /**
@@ -210,46 +154,6 @@ export class Colony {
   }
 
   /**
-   * Create a motion to create a team within a Colony
-   *
-   * For more information about the resulting action, see [[Colony.forceCreateTeam]].
-   *
-   * @remarks
-   * Currently you can only add domains within the `Root` domain. This restriction will be lifted soon
-   *
-   * @param metadataCid An IPFS [CID](https://docs.ipfs.io/concepts/content-addressing/#identifier-formats) for a JSON file containing the metadata described below. For now, we would like to keep it agnostic to any IPFS upload mechanism, so you have to upload the file manually and provide your own hash (by using, for example, [Pinata](https://docs.pinata.cloud/))
-   *
-   * @returns A tupel: `[eventData, ContractReceipt]`
-   *
-   * **Motion event data**
-   * | Property | Type | Description |
-   * | :------ | :------ | :------ |
-   * | `motionId` | BigNumber | ID of the motion created |
-   * | `creator` | string | Address of the motion's creator |
-   * | `domainId` | BigNumber | Team ID of the motion |
-   */
-  async createTeam(metadataCid?: string) {
-    // This will change when we have nested permissions
-    const parentDomain = Id.RootDomain;
-
-    return metadataCid
-      ? this.createMotion(
-          'addDomain(uint256,uint256,uint256,string)',
-          parentDomain,
-          ColonyRole.Architecture,
-          parentDomain,
-          [parentDomain, metadataCid],
-        )
-      : this.createMotion(
-          'addDomain(uint256,uint256,uint256)',
-          parentDomain,
-          ColonyRole.Architecture,
-          parentDomain,
-          [parentDomain],
-        );
-  }
-
-  /**
    * Create a team within a Colony
    *
    * @remarks
@@ -276,7 +180,7 @@ export class Colony {
    * | `domainColor` | string | The color assigned to this team |
    * | `domainPurpose` | string | The purpose for this team (a broad description) |
    */
-  async forceCreateTeam(metadataCid?: string) {
+  async createTeam(metadataCid?: string) {
     let tx: ContractTransaction;
     if (metadataCid) {
       tx = await this.colonyClient['addDomainWithProofs(uint256,string)'](
@@ -340,100 +244,6 @@ export class Colony {
   }
 
   /**
-   * Create a motion to make a payment to a single address using a single token
-   *
-   * For more information about the resulting action, see [[Colony.forcePay]].
-   *
-   * @remarks Requires the `OneTxPayment` extension to be installed for the Colony (this is usually the case for Colonies created via the Dapp). Note that most tokens use 18 decimals, so add a bunch of zeros or use our `w` or `toWei` functions (see example)
-   *
-   * @example
-   * ```typescript
-   * import { Id, Tokens, w } from '@colony/sdk';
-   *
-   * // Immediately executing async function
-   * (async function() {
-   *   // Pay 10 XDAI (on Gnosis chain) from the root domain to the following address
-   *   await colony.pay(
-   *      '0xb77D57F4959eAfA0339424b83FcFaf9c15407461',
-   *      w`10`,
-   *      Id.RootDomain,
-   *      Tokens.Gnosis.XDAI,
-   *   );
-   * })();
-   * ```
-   *
-   * @param recipient Wallet address of account to send the funds to (also awarded reputation when sending the native token)
-   * @param amount Amount to pay in wei
-   * @param tokenAddress The address of the token to make the payment in. Default is the Colony's native token
-   * @param teamId The team to use to send the funds from. Has to have funding of at least the amount you need to send. See [[Colony.moveFundsToTeam]]. Defaults to the Colony's root team
-   * @returns A tupel of event data and contract receipt
-   *
-   * **Motion event data**
-   * | Property | Type | Description |
-   * | :------ | :------ | :------ |
-   * | `motionId` | BigNumber | ID of the motion created |
-   * | `creator` | string | Address of the motion's creator |
-   * | `domainId` | BigNumber | Team ID of the motion |
-   */
-  async pay(
-    recipient: string,
-    amount: BigNumberish,
-    teamId?: BigNumberish,
-    tokenAddress?: string,
-  ) {
-    const setTeamId = teamId || Id.RootDomain;
-    const setTokenAddress =
-      tokenAddress || this.colonyClient.tokenClient.address;
-
-    const oneTxClient = await this.colonyClient.getExtensionClient(
-      Extension.OneTxPayment,
-    );
-
-    // We're not forcing it, create a motion instead
-    const votingReputationClient = await this.getVotingReputationClient();
-
-    const [extensionPDID, extensionCSI] = await getExtensionPermissionProofs(
-      this.colonyClient,
-      setTeamId,
-      oneTxClient.address,
-    );
-    const [userPDID, userCSI] = await getExtensionPermissionProofs(
-      this.colonyClient,
-      setTeamId,
-    );
-
-    const encodedAction = oneTxClient.interface.encodeFunctionData(
-      'makePaymentFundedFromDomain',
-      [
-        extensionPDID,
-        extensionCSI,
-        userPDID,
-        userCSI,
-        [recipient],
-        [setTokenAddress],
-        [amount],
-        setTeamId,
-        // Skill associated with this payment. Ignore for now
-        Id.SkillIgnore,
-      ],
-    );
-
-    const tx = await votingReputationClient.createMotionWithProofs(
-      setTeamId,
-      oneTxClient.address,
-      encodedAction,
-    );
-
-    const receipt = await tx.wait();
-
-    const data = {
-      ...extractEvent<MotionCreatedEventObject>('MotionCreated', receipt),
-    };
-
-    return [data, receipt] as [typeof data, typeof receipt];
-  }
-
-  /**
    * Make a payment to a single address using a single token
    *
    * @remarks Requires the `OneTxPayment` extension to be installed for the Colony (this is usually the case for Colonies created via the Dapp). Note that most tokens use 18 decimals, so add a bunch of zeros or use ethers' `parseUnits` function (see example)
@@ -452,7 +262,7 @@ export class Colony {
    * | `fundamentalId` | BigNumber | The newly added payment id |
    * | `nPayouts` | BigNumber | Number of payouts in total |
    */
-  async forcePay(
+  async pay(
     recipient: string,
     amount: BigNumberish,
     teamId?: BigNumberish,
@@ -482,89 +292,6 @@ export class Colony {
     };
 
     return [data, receipt] as [typeof data, typeof receipt];
-  }
-
-  /**
-   * Create a motion to move funds from one team to another
-   *
-   * For more information about the resulting action, see [[Colony.forceMoveFundsToTeam]].
-   *
-   * After sending funds to and claiming funds for your Colony they will land in a special team, the root team. If you want to make payments from other teams (in order to award reputation in that team) you have to move the funds there first. Use this method to do so.
-   *
-   * @remarks Requires the `Funding` permission in the root team. As soon as teams can be nested, this requires the `Funding` permission in a team that is a parent of both teams in which funds are moved.
-   *
-   * @example
-   * ```typescript
-   * import { utils } from 'ethers';
-   * import { Tokens } from '@colony/sdk';
-   *
-   * // Immediately executing async function
-   * (async function() {
-   *   // Move 10 of the native token from team 2 to team 3
-   *   await colony.moveFundsToTeam(
-   *      utils.parseUnits('10'),
-   *      2,
-   *      3,
-   *   );
-   * })();
-   * ```
-   *
-   * @param amount Amount to transfer between the teams
-   * @param toTeam The team to transfer the funds to
-   * @param fromTeam The team to transfer the funds from. Default is the Root team
-   * @param tokenAddress The address of the token to be transferred. Default is the Colony's native token
-   * @returns A tupel of event data and contract receipt
-   *
-   * **Motion event data**
-   * | Property | Type | Description |
-   * | :------ | :------ | :------ |
-   * | `motionId` | BigNumber | ID of the motion created |
-   * | `creator` | string | Address of the motion's creator |
-   * | `domainId` | BigNumber | Team ID of the motion |
-   */
-  async moveFundsToTeam(
-    amount: BigNumberish,
-    toTeam: BigNumberish,
-    fromTeam?: BigNumberish,
-    tokenAddress?: string,
-  ) {
-    const parentTeam = Id.RootDomain;
-    const setFromTeam = fromTeam || Id.RootDomain;
-    const setTokenAddress =
-      tokenAddress || this.colonyClient.tokenClient.address;
-
-    const { fundingPotId: fromPot } = await this.colonyClient.getDomain(
-      setFromTeam,
-    );
-    const { fundingPotId: toPot } = await this.colonyClient.getDomain(toTeam);
-
-    const fromChildSkillIndex = await getChildIndex(
-      this.colonyClient,
-      parentTeam,
-      setFromTeam,
-    );
-    const toChildSkillIndex = await getChildIndex(
-      this.colonyClient,
-      parentTeam,
-      toTeam,
-    );
-
-    return this.createMotion(
-      // eslint-disable-next-line max-len
-      'moveFundsBetweenPots(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,address)',
-      parentTeam,
-      ColonyRole.Funding,
-      parentTeam,
-      [
-        parentTeam,
-        fromChildSkillIndex,
-        toChildSkillIndex,
-        fromPot,
-        toPot,
-        amount,
-        setTokenAddress,
-      ],
-    );
   }
 
   /**
@@ -605,7 +332,7 @@ export class Colony {
    * | `amount` | BigNumber | The amount that was transferred |
    * | `token` | string | The token address being transferred |
    */
-  async forceMoveFundsToTeam(
+  async moveFundsToTeam(
     amount: BigNumberish,
     toTeam: BigNumberish,
     fromTeam?: BigNumberish,
